Relocate google-sheets-service to server-only api/_lib directory
