+++ netlify/functions/process-form-submissions.ts
@@ -26,7 +26,7 @@ const handler: Handler = async (event: HandlerEvent, context: HandlerContext) =>
 
   try {
     // Import services (dynamic to avoid build issues)
-    const { fetchNewSubmissions } = await import('../../src/lib/services/google-sheets-service');
+    const { fetchNewSubmissions } = await import('../../api/_lib/google-sheets-service');
     const { processFormSubmission, isSubmissionProcessed } = await import('../../src/lib/services/form-processing-service');
 
     console.log('Starting form submission processing...');
