+++ src/lib/services/form-processing-service.ts
@@ -4,12 +4,12 @@
  */
 
 import { supabase } from '../supabase/client';
-import type { FormSubmission } from './google-sheets-service';
+import type { FormSubmission } from '../../../api/_lib/google-sheets-service';
 import {
   parseChildNameAndGrade,
   estimateAgeFromGrade,
   determineTierFromGrade,
-} from './google-sheets-service';
+} from '../../../api/_lib/google-sheets-service';
 
 /**
  * Check if form submission has already been processed
